Extract all Ecto functionality to separate module
This should make it easier to add other adapters in the future and
reduces the need for conditionals due to separation of concerns.
